# openapi.json
{"openapi":"3.1.0","info":{"title":"Job Search Assistant API","description":"Backend for job search assistant platform with AI-powered resume parsing","version":"2.0.0"},"paths":{"/user/register":{"post":{"tags":["User"],"summary":"Register User","operationId":"register_user_user_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/request-registration":{"post":{"tags":["User"],"summary":"Request Registration","operationId":"request_registration_user_request_registration_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/confirm-registration":{"post":{"tags":["User"],"summary":"Confirm Registration","operationId":"confirm_registration_user_confirm_registration_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationVerify"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/login":{"post":{"tags":["User"],"summary":"Login User","operationId":"login_user_user_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_user_user_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/request-password-reset":{"post":{"tags":["User"],"summary":"Request Password Reset","operationId":"request_password_reset_user_request_password_reset_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordResetRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/verify-otp":{"post":{"tags":["User"],"summary":"Verify Otp","operationId":"verify_otp_user_verify_otp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OTPVerify"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/reset-password":{"post":{"tags":["User"],"summary":"Reset Password","operationId":"reset_password_user_reset_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/profile/":{"get":{"tags":["Profile"],"summary":"Get Profile","description":"Get user profile.","operationId":"get_profile_profile__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileOut"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Profile"],"summary":"Create Or Update Profile","description":"Create or update user profile with optional resume processing.","operationId":"create_or_update_profile_profile__post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_or_update_profile_profile__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/profile/upload-resume":{"post":{"tags":["Profile"],"summary":"Upload Resume Only","description":"Upload and process resume independently.","operationId":"upload_resume_only_profile_upload_resume_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_resume_only_profile_upload_resume_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Upload Resume Only Profile Upload Resume Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/profile/resume-status":{"get":{"tags":["Profile"],"summary":"Get Resume Processing Status","description":"Check the status of resume processing.","operationId":"get_resume_processing_status_profile_resume_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Resume Processing Status Profile Resume Status Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/profile/resume":{"delete":{"tags":["Profile"],"summary":"Delete Resume","description":"Delete uploaded resume and parsed data.","operationId":"delete_resume_profile_resume_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/information/":{"get":{"tags":["UserInformation"],"summary":"Get Information","operationId":"get_information_information__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInformationOut"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["UserInformation"],"summary":"Create Or Update Information","operationId":"create_or_update_information_information__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInformationCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInformationOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/information/resume":{"post":{"tags":["UserInformation"],"summary":"Upload Resume","operationId":"upload_resume_information_resume_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_resume_information_resume_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_create_or_update_profile_profile__post":{"properties":{"experiences":{"type":"string","title":"Experiences","default":""},"skills":{"type":"string","title":"Skills","default":""},"projects":{"type":"string","title":"Projects","default":""},"education":{"type":"string","title":"Education","default":""},"courses":{"type":"string","title":"Courses","default":""},"achievements":{"type":"string","title":"Achievements","default":""},"extra_curricular":{"type":"string","title":"Extra Curricular","default":""},"resume":{"anyOf":[{"type":"string","format":"binary"},{"type":"null"}],"title":"Resume"}},"type":"object","title":"Body_create_or_update_profile_profile__post"},"Body_login_user_user_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_user_user_login_post"},"Body_upload_resume_information_resume_post":{"properties":{"resume":{"type":"string","format":"binary","title":"Resume"}},"type":"object","required":["resume"],"title":"Body_upload_resume_information_resume_post"},"Body_upload_resume_only_profile_upload_resume_post":{"properties":{"resume":{"type":"string","format":"binary","title":"Resume"}},"type":"object","required":["resume"],"title":"Body_upload_resume_only_profile_upload_resume_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OTPVerify":{"properties":{"user_id":{"type":"string","format":"email","title":"User Id"},"otp":{"type":"string","title":"Otp"}},"type":"object","required":["user_id","otp"],"title":"OTPVerify"},"PasswordResetRequest":{"properties":{"user_id":{"type":"string","format":"email","title":"User Id"}},"type":"object","required":["user_id"],"title":"PasswordResetRequest"},"PasswordUpdate":{"properties":{"user_id":{"type":"string","format":"email","title":"User Id"},"otp":{"type":"string","title":"Otp"},"password":{"type":"string","title":"Password"}},"type":"object","required":["user_id","otp","password"],"title":"PasswordUpdate"},"RegistrationRequest":{"properties":{"user_id":{"type":"string","format":"email","title":"User Id"},"password":{"type":"string","title":"Password"}},"type":"object","required":["user_id","password"],"title":"RegistrationRequest"},"RegistrationVerify":{"properties":{"user_id":{"type":"string","format":"email","title":"User Id"},"otp":{"type":"string","title":"Otp"}},"type":"object","required":["user_id","otp"],"title":"RegistrationVerify"},"UserCreate":{"properties":{"user_id":{"type":"string","format":"email","title":"User Id"},"password":{"type":"string","title":"Password"}},"type":"object","required":["user_id","password"],"title":"UserCreate"},"UserInformationCreate":{"properties":{"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"street":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Street"},"alternate_email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Alternate Email"},"resume_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Resume Path"}},"type":"object","title":"UserInformationCreate"},"UserInformationOut":{"properties":{"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"street":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Street"},"alternate_email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Alternate Email"},"resume_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Resume Path"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["id"],"title":"UserInformationOut"},"UserOut":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"string","format":"email","title":"User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","user_id","created_at"],"title":"UserOut"},"UserProfileOut":{"properties":{"experiences":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Experiences"},"skills":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Skills"},"projects":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Projects"},"education":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Education"},"courses":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Courses"},"achievements":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Achievements"},"extra_curricular":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Extra Curricular"},"id":{"type":"integer","title":"Id"},"resume_filename":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Resume Filename"},"resume_data":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Resume Data"}},"type":"object","required":["id"],"title":"UserProfileOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/user/login"}}}}}}